at89: split into lib + app
